Fix the use of a template with latest IPython.
The use of a template wasn't working for me with IPython 2.3.1. After some digging in runipy and IPython's codes, I found this solution to fix my issue but I have no idea of what the best fix is. Maybe for compatibility it would ok to set both values, something like
```
                    config=Config({'HTMLExporter': {
                        'default_template': args.template,
                        'template_file': args.template,
                    }}))
```