Change nbformat call.
IPython.nbformat.current has been deprecated.

The arguments passed to function `read` changed in IPython 3.0. It now
asks for the notebook version. The new version is v4.

If the user is using IPython < 3, it loads the notebook as before. If
it is 3.0 or greater, it will load the notebook as format v3. I tried to
load as v4, but its format different and more work is needed.